requirements.txt
Django==5.2.8
Pillow==12.0.0
sorl-thumbnail==12.9.0
